requirements.txt
ffmpeg-python
matplotlib
numba
numpy
opencv-python
opencv-contrib-python
plotly
requests
scikit-image